feat: added k3s metrics mappings for apps
This change adds mappings to the graphite exporter configuration file that
map the k3s metrics into unified prometheus metrics. When running apps in
TrueNAS scale the pods are exported by the metrics exporter and can be
captured here to get some basic operational details.

The metrics exported are not documented in TrueNAS or k3s in this format, and
this poses a problem making sense of the `cpu` metric.