feat: k3s pod dashboard
This dashboard gives an overview of the k3s cluster as whole and a collapsable
and repeatable section for each pod.

The cpu metric was identified to be instantaneous cpu time in ns for a given
second. This makes the metric a bit tricky to work with as it does not play
nice with graphana/prometheus' rate intervals, but each value can be computed on
the whole by dividing it by 1bn.